[mobile a#15] 학교 내의 나의 랭크 VO 생성

+++ app/src/main/java/com/opgg/chai/model/data/RankItem.kt
@@ -1,6 +1,6 @@
 package com.opgg.chai.model.data
 
-data class RankItem(val id: Long = 0,
+data class RankItem(val id: Int = 0,
                     val name: String = "",
                     val score: String = "",
                     val isRankUp: Boolean = false,
